fix interface code.

+++ services/ims_service_interaction/src/ims_call_callback_stub.cpp
@@ -43,58 +43,80 @@ void ImsCallCallbackStub::InitFuncMap()
 void ImsCallCallbackStub::InitCallBasicFuncMap()
 {
     /****************** call basic ******************/
-    requestFuncMap_[ImsCallCallbackInterfaceCode::IMS_DIAL] = &ImsCallCallbackStub::OnDialResponseInner;
-    requestFuncMap_[ImsCallCallbackInterfaceCode::IMS_HANG_UP] = &ImsCallCallbackStub::OnHangUpResponseInner;
-    requestFuncMap_[ImsCallCallbackInterfaceCode::IMS_REJECT] = &ImsCallCallbackStub::OnRejectResponseInner;
-    requestFuncMap_[ImsCallCallbackInterfaceCode::IMS_ANSWER] = &ImsCallCallbackStub::OnAnswerResponseInner;
-    requestFuncMap_[ImsCallCallbackInterfaceCode::IMS_HOLD] = &ImsCallCallbackStub::OnHoldCallResponseInner;
-    requestFuncMap_[ImsCallCallbackInterfaceCode::IMS_UN_HOLD] = &ImsCallCallbackStub::OnUnHoldCallResponseInner;
-    requestFuncMap_[ImsCallCallbackInterfaceCode::IMS_SWITCH] = &ImsCallCallbackStub::OnSwitchCallResponseInner;
-    requestFuncMap_[ImsCallCallbackInterfaceCode::IMS_CALL_STATE_CHANGE] =
+    requestFuncMap_[static_cast<uint32_t>(ImsCallCallbackInterfaceCode::IMS_DIAL)] =
+        &ImsCallCallbackStub::OnDialResponseInner;
+    requestFuncMap_[static_cast<uint32_t>(ImsCallCallbackInterfaceCode::IMS_HANG_UP)] =
+        &ImsCallCallbackStub::OnHangUpResponseInner;
+    requestFuncMap_[static_cast<uint32_t>(ImsCallCallbackInterfaceCode::IMS_REJECT)] =
+        &ImsCallCallbackStub::OnRejectResponseInner;
+    requestFuncMap_[static_cast<uint32_t>(ImsCallCallbackInterfaceCode::IMS_ANSWER)] =
+        &ImsCallCallbackStub::OnAnswerResponseInner;
+    requestFuncMap_[static_cast<uint32_t>(ImsCallCallbackInterfaceCode::IMS_HOLD)] =
+        &ImsCallCallbackStub::OnHoldCallResponseInner;
+    requestFuncMap_[static_cast<uint32_t>(ImsCallCallbackInterfaceCode::IMS_UN_HOLD)] =
+        &ImsCallCallbackStub::OnUnHoldCallResponseInner;
+    requestFuncMap_[static_cast<uint32_t>(ImsCallCallbackInterfaceCode::IMS_SWITCH)] =
+        &ImsCallCallbackStub::OnSwitchCallResponseInner;
+    requestFuncMap_[static_cast<uint32_t>(ImsCallCallbackInterfaceCode::IMS_CALL_STATE_CHANGE)] =
         &ImsCallCallbackStub::OnCallStateChangeReportInner;
-    requestFuncMap_[ImsCallCallbackInterfaceCode::IMS_LAST_CALL_FAIL_REASON] =
+    requestFuncMap_[static_cast<uint32_t>(ImsCallCallbackInterfaceCode::IMS_LAST_CALL_FAIL_REASON)] =
         &ImsCallCallbackStub::OnLastCallFailReasonResponseInner;
-    requestFuncMap_[ImsCallCallbackInterfaceCode::IMS_CALL_CRING] = &ImsCallCallbackStub::OnCallRingBackReportInner;
+    requestFuncMap_[static_cast<uint32_t>(ImsCallCallbackInterfaceCode::IMS_CALL_CRING)] =
+        &ImsCallCallbackStub::OnCallRingBackReportInner;
 }
 
 void ImsCallCallbackStub::InitConfigFuncMap()
 {
     /****************** dtmf rtt ******************/
-    requestFuncMap_[ImsCallCallbackInterfaceCode::IMS_START_DTMF] = &ImsCallCallbackStub::OnStartDtmfResponseInner;
-    requestFuncMap_[ImsCallCallbackInterfaceCode::IMS_SEND_DTMF] = &ImsCallCallbackStub::OnSendDtmfResponseInner;
-    requestFuncMap_[ImsCallCallbackInterfaceCode::IMS_STOP_DTMF] = &ImsCallCallbackStub::OnStopDtmfResponseInner;
+    requestFuncMap_[static_cast<uint32_t>(ImsCallCallbackInterfaceCode::IMS_START_DTMF)] =
+        &ImsCallCallbackStub::OnStartDtmfResponseInner;
+    requestFuncMap_[static_cast<uint32_t>(ImsCallCallbackInterfaceCode::IMS_SEND_DTMF)] =
+        &ImsCallCallbackStub::OnSendDtmfResponseInner;
+    requestFuncMap_[static_cast<uint32_t>(ImsCallCallbackInterfaceCode::IMS_STOP_DTMF)] =
+        &ImsCallCallbackStub::OnStopDtmfResponseInner;
 
     /****************** ims config ******************/
-    requestFuncMap_[ImsCallCallbackInterfaceCode::IMS_SET_SWITCH_STATUS] =
+    requestFuncMap_[static_cast<uint32_t>(ImsCallCallbackInterfaceCode::IMS_SET_SWITCH_STATUS)] =
         &ImsCallCallbackStub::OnSetImsSwitchResponseInner;
-    requestFuncMap_[ImsCallCallbackInterfaceCode::IMS_GET_SWITCH_STATUS] =
+    requestFuncMap_[static_cast<uint32_t>(ImsCallCallbackInterfaceCode::IMS_GET_SWITCH_STATUS)] =
         &ImsCallCallbackStub::OnGetImsSwitchResponseInner;
-    requestFuncMap_[ImsCallCallbackInterfaceCode::IMS_GET_CALLS_DATA] =
+    requestFuncMap_[static_cast<uint32_t>(ImsCallCallbackInterfaceCode::IMS_GET_CALLS_DATA)] =
         &ImsCallCallbackStub::OnGetImsCallsDataResponseInner;
-    requestFuncMap_[ImsCallCallbackInterfaceCode::IMS_SET_MUTE] = &ImsCallCallbackStub::OnSetMuteResponseInner;
+    requestFuncMap_[static_cast<uint32_t>(ImsCallCallbackInterfaceCode::IMS_SET_MUTE)] =
+        &ImsCallCallbackStub::OnSetMuteResponseInner;
 }
 
 void ImsCallCallbackStub::InitSupplementFuncMap()
 {
     /****************** supplement ******************/
-    requestFuncMap_[ImsCallCallbackInterfaceCode::IMS_SET_CALL_CLIP] = &ImsCallCallbackStub::OnSetClipResponseInner;
-    requestFuncMap_[ImsCallCallbackInterfaceCode::IMS_GET_CALL_CLIP] = &ImsCallCallbackStub::OnGetClipResponseInner;
-    requestFuncMap_[ImsCallCallbackInterfaceCode::IMS_GET_CALL_CLIR] = &ImsCallCallbackStub::OnGetClirResponseInner;
-    requestFuncMap_[ImsCallCallbackInterfaceCode::IMS_SET_CALL_CLIR] = &ImsCallCallbackStub::OnSetClirResponseInner;
-    requestFuncMap_[ImsCallCallbackInterfaceCode::IMS_GET_CALL_FORWARD] =
+    requestFuncMap_[static_cast<uint32_t>(ImsCallCallbackInterfaceCode::IMS_SET_CALL_CLIP)] =
+        &ImsCallCallbackStub::OnSetClipResponseInner;
+    requestFuncMap_[static_cast<uint32_t>(ImsCallCallbackInterfaceCode::IMS_GET_CALL_CLIP)] =
+        &ImsCallCallbackStub::OnGetClipResponseInner;
+    requestFuncMap_[static_cast<uint32_t>(ImsCallCallbackInterfaceCode::IMS_GET_CALL_CLIR)] =
+        &ImsCallCallbackStub::OnGetClirResponseInner;
+    requestFuncMap_[static_cast<uint32_t>(ImsCallCallbackInterfaceCode::IMS_SET_CALL_CLIR)] =
+        &ImsCallCallbackStub::OnSetClirResponseInner;
+    requestFuncMap_[static_cast<uint32_t>(ImsCallCallbackInterfaceCode::IMS_GET_CALL_FORWARD)] =
         &ImsCallCallbackStub::OnGetCallTransferResponseInner;
-    requestFuncMap_[ImsCallCallbackInterfaceCode::IMS_SET_CALL_FORWARD] =
+    requestFuncMap_[static_cast<uint32_t>(ImsCallCallbackInterfaceCode::IMS_SET_CALL_FORWARD)] =
         &ImsCallCallbackStub::OnSetCallTransferResponseInner;
-    requestFuncMap_[ImsCallCallbackInterfaceCode::IMS_GET_CALL_RESTRICTION] =
+    requestFuncMap_[static_cast<uint32_t>(ImsCallCallbackInterfaceCode::IMS_GET_CALL_RESTRICTION)] =
         &ImsCallCallbackStub::OnGetCallRestrictionResponseInner;
-    requestFuncMap_[ImsCallCallbackInterfaceCode::IMS_SET_CALL_RESTRICTION] =
+    requestFuncMap_[static_cast<uint32_t>(ImsCallCallbackInterfaceCode::IMS_SET_CALL_RESTRICTION)] =
         &ImsCallCallbackStub::OnSetCallRestrictionResponseInner;
-    requestFuncMap_[ImsCallCallbackInterfaceCode::IMS_GET_CALL_WAIT] = &ImsCallCallbackStub::OnGetCallWaitingResponseInner;
-    requestFuncMap_[ImsCallCallbackInterfaceCode::IMS_SET_CALL_WAIT] = &ImsCallCallbackStub::OnSetCallWaitingResponseInner;
-    requestFuncMap_[ImsCallCallbackInterfaceCode::IMS_GET_CALL_COLR] = &ImsCallCallbackStub::OnGetColrResponseInner;
-    requestFuncMap_[ImsCallCallbackInterfaceCode::IMS_SET_CALL_COLR] = &ImsCallCallbackStub::OnSetColrResponseInner;
-    requestFuncMap_[ImsCallCallbackInterfaceCode::IMS_SET_CALL_COLP] = &ImsCallCallbackStub::OnSetColpResponseInner;
-    requestFuncMap_[ImsCallCallbackInterfaceCode::IMS_GET_CALL_COLP] = &ImsCallCallbackStub::OnGetColpResponseInner;
+    requestFuncMap_[static_cast<uint32_t>(ImsCallCallbackInterfaceCode::IMS_GET_CALL_WAIT)] =
+        &ImsCallCallbackStub::OnGetCallWaitingResponseInner;
+    requestFuncMap_[static_cast<uint32_t>(ImsCallCallbackInterfaceCode::IMS_SET_CALL_WAIT)] =
+        &ImsCallCallbackStub::OnSetCallWaitingResponseInner;
+    requestFuncMap_[static_cast<uint32_t>(ImsCallCallbackInterfaceCode::IMS_GET_CALL_COLR)] =
+        &ImsCallCallbackStub::OnGetColrResponseInner;
+    requestFuncMap_[static_cast<uint32_t>(ImsCallCallbackInterfaceCode::IMS_SET_CALL_COLR)] =
+        &ImsCallCallbackStub::OnSetColrResponseInner;
+    requestFuncMap_[static_cast<uint32_t>(ImsCallCallbackInterfaceCode::IMS_SET_CALL_COLP)] =
+        &ImsCallCallbackStub::OnSetColpResponseInner;
+    requestFuncMap_[static_cast<uint32_t>(ImsCallCallbackInterfaceCode::IMS_GET_CALL_COLP)] =
+        &ImsCallCallbackStub::OnGetColpResponseInner;
 }
 
 ImsCallCallbackStub::~ImsCallCallbackStub()
